@@ -10,7 +10,7 @@ set(OPENSMT_VERSION_PATCH 1)
 set(OPENSMT_VERSION ${OPENSMT_VERSION_MAJOR}.${OPENSMT_VERSION_MINOR}.${OPENSMT_VERSION_PATCH})
 project(opensmt VERSION ${OPENSMT_VERSION} LANGUAGES CXX)
 
-set(CMAKE_CXX_STANDARD 11)
+set(CMAKE_CXX_STANDARD 17)
 
 set(CMAKE_SOURCE_DIR "src")
 
